src/main/java/com/challange/orderjavachallange/user/mapper/UserMapper.java
package com.challange.orderjavachallange.user.mapper;

import com.challange.orderjavachallange.user.dto.UserDto;
import com.challange.orderjavachallange.user.entity.User;

public class UserMapper {

    public static UserDto toDto(User user) {
        if (user == null) {
            return null;
        }
        return new UserDto(
                user.getId(),
                user.getName(),
                user.getEmail()
        );
    }

    public static User toEntity(UserDto userDto) {
        if (userDto == null) {
            return null;
        }
        User user = new User();
        user.setId(userDto.getId());
        user.setName(userDto.getName());
        user.setEmail(userDto.getEmail());
        return user;
    }
}
